Move cluster to init
